Zutat löschen blockieren wenn verwendet

+++ app/src/main/java/com/example/eaeprojekt/popups/PopupEditIngredient.java
@@ -3,20 +3,17 @@ package com.example.eaeprojekt.popups;
 
 import android.app.Activity;
 import android.util.Log;
-import android.view.Gravity;
-import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.EditText;
-import android.widget.FrameLayout;
-import android.widget.LinearLayout;
-import android.widget.PopupWindow;
-import android.widget.Spinner;
+import android.widget.Toast;
 
 import androidx.constraintlayout.widget.ConstraintLayout;
 
+import com.example.eaeprojekt.DTO.IngredientAmountDTO;
 import com.example.eaeprojekt.R;
-import com.example.eaeprojekt.activity.NewRecipeActivity;
 import com.example.eaeprojekt.database.DatabaseManager;
+
+import java.util.List;
 
 public class PopupEditIngredient implements View.OnClickListener {
 
@@ -26,7 +23,6 @@ public class PopupEditIngredient implements View.OnClickListener {
     ConstraintLayout buttonEdit;
     ConstraintLayout buttonBack;
     ConstraintLayout buttonDelete;
-    FrameLayout frame;
     Activity mainActivity;
 
     long id;
@@ -65,8 +61,8 @@ public class PopupEditIngredient implements View.OnClickListener {
             db = new DatabaseManager(mainActivity);
             db.open();
 
-            EditText name = (EditText) parentView.findViewById(R.id.nameText2);
-            EditText unit = (EditText) parentView.findViewById(R.id.unitText2);
+            EditText name = parentView.findViewById(R.id.nameText2);
+            EditText unit = parentView.findViewById(R.id.unitText2);
 
             db.updateIngredient(id, name.getText().toString(), unit.getText().toString());
 
@@ -78,8 +74,21 @@ public class PopupEditIngredient implements View.OnClickListener {
         } else if (viewClick == buttonDelete) {
             db = new DatabaseManager(mainActivity);
             db.open();
-
+            boolean isUsed = false;
+            List<IngredientAmountDTO> allIngredientAmounts = db.getAllIngredientAmounts();
+            for (IngredientAmountDTO i : allIngredientAmounts) {
+                if (i.getIngredientId() == id) {
+                    isUsed = true;
+                    break;
+                }
+            }
+            if (!isUsed)
             db.deleteIngredient(id);
+            else {
+                Toast toast = new Toast(mainActivity);
+                toast.setText(R.string.ingredientAmountCannotBeDeleted);
+                toast.show();
+            }
 
             parentView.findViewById(R.id.above).setVisibility(View.GONE);
             parentView.findViewById(R.id.bbelow).setVisibility(View.VISIBLE);
